requirements.txt
streamlit
requests
openai
python-dotenv
SpeechRecognition
pyaudio
